Activate Nullable; improve Domain layer code

+++ src/MusicManager/MusicManager.Domain.Test/UnitTesting/TestHelper.cs
@@ -12,33 +12,28 @@ namespace Test.MusicManager.Domain.UnitTesting
     [TestClass]
     public static class TestHelper
     {
-        private static HashSet<string> tempFiles = new HashSet<string>();
+        private static readonly HashSet<string> tempFiles = new();
         
-        [AssemblyInitialize]
-        public static void Initialize(TestContext testContext)
-        {
-        }
-
-        public static string GetTempFileName(string extension = null)
+        public static string GetTempFileName(string? extension = null)
         {
             var tempFile = Path.Combine(Path.GetTempPath(), "tmp" + Path.GetRandomFileName());
-            if (!string.IsNullOrEmpty(extension)) { tempFile += extension; }
+            if (!string.IsNullOrEmpty(extension)) tempFile += extension;
             tempFiles.Add(tempFile);
             return tempFile;
         }
 
-        public static void AssertHaveEqualPropertyValues<T>(T expected, T actual, Func<PropertyInfo, bool> predicate = null)
+        public static void AssertHaveEqualPropertyValues<T>(T expected, T actual, Func<PropertyInfo, bool>? predicate = null)
         {
             var objectType = typeof(T);
             var properties = objectType.GetProperties();
-            predicate = predicate ?? (p => true);
+            predicate ??= (p => true);
 
             foreach (var property in properties.Where(predicate))
             {
                 if (property.PropertyType != typeof(string) && typeof(IEnumerable).IsAssignableFrom(property.PropertyType))
                 {
-                    var expectedCollection = ((IEnumerable)property.GetValue(expected)).Cast<object>();
-                    var actualCollection = ((IEnumerable)property.GetValue(actual)).Cast<object>();
+                    var expectedCollection = ((IEnumerable)property.GetValue(expected)!).Cast<object>();
+                    var actualCollection = ((IEnumerable)property.GetValue(actual)!).Cast<object>();
                     AssertHelper.SequenceEqual(expectedCollection, actualCollection);
                 }
                 else
@@ -53,10 +48,7 @@ namespace Test.MusicManager.Domain.UnitTesting
         {
             foreach (var tempFile in tempFiles)
             {
-                if (File.Exists(tempFile))
-                {
-                    File.Delete(tempFile);
-                }
+                if (File.Exists(tempFile)) File.Delete(tempFile);
             }
         }
     }
